feat: create user node creation gRPC method (#62)
* feat: set gradle project

* feat: create createUser interface

* feat: add generated ignore

* feat: import common-module

* rename: rename directory (pallete->palette)

* feat: create gRPC user create service

* feat: add eureka setting and ports

* feat: add real business the social user create method

* feat: add gRPC config
