+++ src/common-module/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Thu Jan 18 15:05:45 KST 2024
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-8.4-bin.zip
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
