+++ src/social-service/settings.gradle.kts
@@ -1,1 +1,4 @@
 rootProject.name = "easel-social-service"
+
+include(":common-module")
+findProject(":common-module")?.projectDir = file("../common-module")
