+++ src/social-service/src/main/java/org/palette/easelsocialservice/EaselSocialServiceApplication.java
@@ -1,0 +1,15 @@
+package org.palette.easelsocialservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@EnableDiscoveryClient
+@SpringBootApplication
+public class EaselSocialServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(EaselSocialServiceApplication.class, args);
+	}
+
+}
