+++ src/social-service/src/main/java/org/palette/easelsocialservice/persistence/HashtagRepository.java
@@ -1,0 +1,6 @@
+package org.palette.easelsocialservice.persistence;
+
+import org.palette.easelsocialservice.persistence.domain.Hashtag;
+import org.springframework.data.neo4j.repository.Neo4jRepository;
+public interface HashtagRepository extends Neo4jRepository<Hashtag, String> {
+}
