+++ src/social-service/src/main/java/org/palette/easelsocialservice/persistence/PaintRepository.java
@@ -1,0 +1,11 @@
+package org.palette.easelsocialservice.persistence;
+
+import org.palette.easelsocialservice.persistence.domain.Paint;
+import org.springframework.data.neo4j.repository.Neo4jRepository;
+
+import java.util.Optional;
+
+public interface PaintRepository extends Neo4jRepository<Paint, Long> {
+    Optional<Paint> findByPid(Long pid);
+
+}
