+++ src/social-service/src/main/java/org/palette/easelsocialservice/persistence/domain/Tags.java
@@ -1,0 +1,24 @@
+package org.palette.easelsocialservice.persistence.domain;
+
+import org.springframework.data.neo4j.core.schema.RelationshipId;
+import org.springframework.data.neo4j.core.schema.RelationshipProperties;
+import org.springframework.data.neo4j.core.schema.TargetNode;
+
+@RelationshipProperties
+public class Tags {
+    @RelationshipId
+    private Long id;
+
+    @TargetNode
+    private Hashtag hashtag;
+
+    private Integer startIdx;
+
+    private Integer endIdx;
+
+    public Tags(Hashtag hashtag, Integer startIdx, Integer endIdx) {
+        this.hashtag = hashtag;
+        this.startIdx = startIdx;
+        this.endIdx = endIdx;
+    }
+}
