+++ src/social-service/src/main/java/org/palette/easelsocialservice/persistence/domain/Uses.java
@@ -1,0 +1,18 @@
+package org.palette.easelsocialservice.persistence.domain;
+
+import org.springframework.data.neo4j.core.schema.RelationshipId;
+import org.springframework.data.neo4j.core.schema.RelationshipProperties;
+import org.springframework.data.neo4j.core.schema.TargetNode;
+
+@RelationshipProperties
+public class Uses {
+    @RelationshipId
+    private Long id;
+
+    @TargetNode
+    private Media media;
+
+    public Uses(Media media) {
+        this.media = media;
+    }
+}
