+++ src/social-service/src/main/java/org/palette/easelsocialservice/service/PaintService.java
@@ -1,0 +1,12 @@
+package org.palette.easelsocialservice.service;
+
+
+import io.grpc.stub.StreamObserver;
+import net.devh.boot.grpc.server.service.GrpcService;
+import org.palette.grpc.GCreateUserRequest;
+import org.palette.grpc.GCreateUserResponse;
+import org.palette.grpc.GSocialServiceGrpc;
+
+public class PaintService {
+
+}
