--- src/social-service/src/main/java/org/pallete/easelsocialservice/EaselSocialServiceApplication.java
@@ -1,13 +1,0 @@
-package org.pallete.easelsocialservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class EaselSocialServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(EaselSocialServiceApplication.class, args);
-	}
-
-}
