--- src/social-service/src/main/java/org/pallete/easelsocialservice/config/Neo4jConfig.java
@@ -1,15 +1,0 @@
-package org.pallete.easelsocialservice.config;
-
-import org.springframework.context.annotation.Bean;
-import org.neo4j.cypherdsl.core.renderer.Configuration;
-import org.neo4j.cypherdsl.core.renderer.Dialect;
-
-@org.springframework.context.annotation.Configuration
-public class Neo4jConfig {
-
-    @Bean
-    Configuration cypherDslConfiguration() {
-        return Configuration.newConfig().withDialect(Dialect.NEO4J_5).build();
-    }
-}
-
