--- src/social-service/src/main/java/org/pallete/easelsocialservice/persistence/LinkRepository.java
@@ -1,7 +1,0 @@
-package org.pallete.easelsocialservice.persistence;
-
-import org.pallete.easelsocialservice.persistence.domain.Link;
-import org.springframework.data.neo4j.repository.Neo4jRepository;
-
-public interface LinkRepository extends Neo4jRepository<Link, String> {
-}
