--- src/social-service/src/main/java/org/pallete/easelsocialservice/persistence/domain/Link.java
@@ -1,17 +1,0 @@
-package org.pallete.easelsocialservice.persistence.domain;
-
-import org.springframework.data.neo4j.core.schema.Id;
-import org.springframework.data.neo4j.core.schema.Node;
-
-@Node("link")
-public class Link {
-    @Id
-    private String shortUrl;
-
-    private String originalUrl;
-
-    public Link(String shortUrl, String originalUrl) {
-        this.shortUrl = shortUrl;
-        this.originalUrl = originalUrl;
-    }
-}
