--- src/social-service/src/main/java/org/pallete/easelsocialservice/persistence/domain/Media.java
@@ -1,17 +1,0 @@
-package org.pallete.easelsocialservice.persistence.domain;
-
-import org.springframework.data.neo4j.core.schema.Id;
-import org.springframework.data.neo4j.core.schema.Node;
-
-@Node("media")
-public class Media {
-    @Id
-    private String path;
-
-    private String type;
-
-    public Media(String path, String type) {
-        this.path = path;
-        this.type = type;
-    }
-}
