+++ src/social-service/src/main/resources/application.yml
@@ -1,3 +1,6 @@
+server:
+  port: 10003
+
 spring:
   application:
     name: social-service
@@ -7,3 +10,16 @@ spring:
       username: ${SOCIAL_DB_USERNAME}
       password: ${SOCIAL_DB_PASSWORD}
     uri: ${SOCIAL_DB_URI}
+
+eureka:
+  instance:
+    hostname: localhost
+  client:
+    fetch-registry: true
+    register-with-eureka: true
+    service-url:
+      defaultZone: http://localhost:8761/eureka
+
+grpc:
+  server:
+    port: 11003
