+++ src/social-service/src/test/java/org/palette/easelsocialservice/EaselSocialServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package org.palette.easelsocialservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EaselSocialServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
