--- src/social-service/src/test/java/org/pallete/easelsocialservice/EaselSocialServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package org.pallete.easelsocialservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EaselSocialServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
